requirements.txt
langchain
langchain_community
openai
numpy
tqdm
google-generativeai
langchain_google_genai